FUNr/requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.19.0
Flask>=2.0.0